+++ NEWS.md
@@ -16,6 +16,7 @@ v3.0.0-a0
 - Drop support for rarfiles.
 - optiping optimization level changed to o5
 - Exif & icc metadata preserved in all formats much better by default.
+- Fall back to Pillow optimization if no external programs.
 
 v2.2.1